changed double to float

+++ inc/kin_math.h
@@ -26,30 +26,30 @@ extern "C" {
 typedef struct {
 	size_t rows;
 	size_t cols;
-	double *data;
+	float *data;
 } matrix_t;
 
 typedef struct {
 	matrix_t *vector;
-	double value;
+	float value;
 } eigen_t;
 
 void free_matrix(matrix_t *matrix);
 matrix_t *init_matrix(size_t rows, size_t cols);
-matrix_t *arr_to_matrix(double *arr, size_t rows, size_t cols);
+matrix_t *arr_to_matrix(float *arr, size_t rows, size_t cols);
 void print_matrix(const matrix_t *matrix);
-void print_arr(const double *arr, size_t size);
+void print_arr(const float *arr, size_t size);
 
 matrix_t *add_matrix(const matrix_t *a, const matrix_t *b);
 matrix_t *sub_matrix(const matrix_t *a, const matrix_t *b);
 matrix_t *mul_matrix(const matrix_t *a, const matrix_t *b);
-matrix_t *scale_matrix(matrix_t *matrix, double scalar);
+matrix_t *scale_matrix(matrix_t *matrix, float scalar);
 
 matrix_t *trans_matrix(const matrix_t *matrix);
 matrix_t *ident_matrix(size_t size);
 
 matrix_t *inv_matrix(matrix_t *matrix);
-double matrix_norm(const matrix_t *matrix);
+float matrix_norm(const matrix_t *matrix);
 matrix_t *normalize_matrix(matrix_t *matrix);
 
 eigen_t *eigen_matrix(const matrix_t *matrix_t);
@@ -61,11 +61,11 @@ matrix_t *rot_matrix_to_quat(matrix_t *matrix);
 
 matrix_t *mul_quat(const matrix_t *a, const matrix_t *b);
 matrix_t *mul_vector(const matrix_t *a, const matrix_t *b);
-double vector_dot(const matrix_t *a, const matrix_t *b);
+float vector_dot(const matrix_t *a, const matrix_t *b);
 
-int sgn(double x);
-double rad_to_deg(double rad);
-double deg_to_rad(double deg);
+int sgn(float x);
+float rad_to_deg(float rad);
+float deg_to_rad(float deg);
 
 #ifdef __cplusplus
 }
